requirements.txt
streamlit
openai
pinecone
python-dotenv
google-generativeai
